+++ src/game/game.c
@@ -25,12 +25,24 @@ bool StartGame(int32_t level_num, GAMEFLOW_LEVEL_TYPE level_type)
 {
     g_CurrentLevel = level_num;
     g_GameInfo.current_level_type = level_type;
-    if (level_type == GFL_SAVED) {
+    if (level_type == GFL_SAVED || level_type == GFL_RESTART) {
         // reset start info to the defaults so that we do not do
         // GlobalItemReplace in the inventory initialization routines too early
         Savegame_ResetStartInfo(level_num);
     } else {
         InitialiseLevelFlags();
+    }
+
+    if (level_type == GFL_RESTART) {
+        int16_t savegame_level_num =
+            Savegame_GetLevelNumber(g_InvExtraData[IED_SAVEGAME_NUM]);
+        if (!InitialiseLevel(savegame_level_num)) {
+            return false;
+        }
+        if (!Savegame_Load(g_InvExtraData[IED_SAVEGAME_NUM], &g_GameInfo)) {
+            LOG_ERROR("Failed to load save file!");
+            return false;
+        }
     }
 
     if (!InitialiseLevel(level_num)) {
@@ -79,9 +91,7 @@ int32_t StopGame(void)
     } else if (
         g_InvExtraData[IED_PAGE_NUM] == PASSPORT_PAGE_1
         && g_InvExtraData[IED_PASSPORT_MODE] == PASSPORT_MODE_SELECT_LEVEL) {
-        // TODO Placeholder for select level. Do new game.
-        Savegame_InitStartCurrentInfo();
-        return GF_START_GAME | g_InvExtraData[IED_LEVEL_NUM];
+        return GF_RESTART_GAME | g_InvExtraData[IED_LEVEL_NUM];
     } else if (g_InvExtraData[IED_PAGE_NUM] == PASSPORT_PAGE_2) {
         return GF_START_GAME
             | (g_InvMode == INV_DEATH_MODE ? g_CurrentLevel