+++ src/game/inventry.c
@@ -691,11 +691,8 @@ int32_t Display_Inventory(int inv_mode)
                 && g_InvExtraData[IED_PASSPORT_MODE]
                     == PASSPORT_MODE_SELECT_LEVEL) {
                 // page 1: select level
-                // TODO Placeholder for select level. Do new game.
-                Savegame_InitStartCurrentInfo();
-                return GF_START_GAME | g_InvExtraData[IED_LEVEL_NUM];
+                return GF_RESTART_GAME | g_InvExtraData[IED_LEVEL_NUM];
             } else if (g_InvExtraData[IED_PAGE_NUM] == PASSPORT_PAGE_2) {
-                // page 2: new game
                 switch (g_InvExtraData[IED_SAVEGAME_NUM]) {
                 case 0:
                     g_GameInfo.bonus_flag = 0;
@@ -727,9 +724,7 @@ int32_t Display_Inventory(int inv_mode)
                 && g_InvExtraData[IED_PASSPORT_MODE]
                     == PASSPORT_MODE_SELECT_LEVEL) {
                 // page 1: select level
-                // TODO Placeholder for select level. Do new game.
-                Savegame_InitStartCurrentInfo();
-                return GF_START_GAME | g_InvExtraData[IED_LEVEL_NUM];
+                return GF_RESTART_GAME | g_InvExtraData[IED_LEVEL_NUM];
             } else if (g_InvExtraData[IED_PAGE_NUM] == PASSPORT_PAGE_2) {
                 // page 2: restart level
                 return GF_RESTART_GAME | g_CurrentLevel;
@@ -748,9 +743,7 @@ int32_t Display_Inventory(int inv_mode)
                 && g_InvExtraData[IED_PASSPORT_MODE]
                     == PASSPORT_MODE_SELECT_LEVEL) {
                 // page 1: select level
-                // TODO Placeholder for select level. Do new game.
-                Savegame_InitStartCurrentInfo();
-                return GF_START_GAME | g_InvExtraData[IED_LEVEL_NUM];
+                return GF_RESTART_GAME | g_InvExtraData[IED_LEVEL_NUM];
             } else if (g_InvExtraData[IED_PAGE_NUM] == PASSPORT_PAGE_2) {
                 // page 1: save game, or new game in gym
                 if (g_CurrentLevel == g_GameFlow.gym_level_num) {